pyproject.toml
[tool.poetry]
name = "projeto-04-04-datapipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "projeto_04"}]

[tool.poetry.dependencies]
python = "3.12.*"
ipykernel = "^6.29.5"
newsapi-python = "^0.2.7"
pandas = "^2.2.3"
numpy = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
